pyproject.toml
[project]
name = "fashion-recommender-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <=3.13.2"    
dependencies = [
    "datasets>=3.5.0",
    "litellm>=1.65.0",
    "matplotlib>=3.10.1",
    "pillow>=11.1.0",
    "pydantic-settings>=2.8.1",
    "qdrant-client>=1.13.3",
    "sentence-transformers>=4.0.1",
]
